# microoptimization on rand(::AliasTable)