--- a/src/samplers/aliastable.jl
+++ b/src/samplers/aliastable.jl
@@ -15,6 +15,7 @@ end
 
 function rand(rng::AbstractRNG, s::AliasTable)
     i = rand(rng, 1:length(s.alias)) % Int
+    # using `ifelse` improves performance here: github.com/JuliaStats/Distributions.jl/pull/1831/
     ifelse(rand(rng) < s.accept[i],  i, s.alias[i])
 end