@devmotion 2024-01-31
From the docs of `ifelse`:

```md
In some cases, using `ifelse` instead of an `if` statement can eliminate the branch in generated code and provide higher performance in tight loops.
```

Maybe something along these lines is the reason for the performance improvements? Did you compare the output of `@code_llvm`?