@adienes 2024-01-31
that's certainly plausible. here is the output of `@code_llvm` on my computer. left is status quo, right is PR

I do not really know how to interpret it though :) 

![image](https://github.com/JuliaStats/Distributions.jl/assets/51664769/82a72e4b-3a6f-4c4a-b315-b2b3ce97039d)