@Zentrik 2024-02-01
It looks like the `ifelse` does remove a branch in the LLVM IR however the lack of `@inbounds` adds some branching. I suspect `ifelse` and `@inbounds` would produce the fastest result.