@adienes 2024-02-01
I did try with `@inbounds` here and it is actually ever-so-slightly slower (but basically equivalent) on my machine. Given the performance equivalence, I defaulted to the "safer" option.

however, I don't know if that result is cpu-specific or not though. is it possible `@inbounds` would have a bigger impact on a different architecture?