@Zentrik 2024-02-01
That seems plausible to me. I imagine the CPU is able to eliminate the cost of bounds checking whilst benchmarking due to branch prediction.