@adienes 2024-02-01
makes sense. happy to add back in the `@inbounds`

anything else needed? I can do more careful benchmarking across a range of inputs, but tbh this seems like a pretty quick win