@devmotion 2024-02-01
IMO we should only use `@inbounds` if it is checked that the indices are valid - but as the function is written currently `i` is only guaranteed to be a valid index of `s.alias`, not of `s.accept`.