@devmotion 2024-02-01
It would definitely be less likely to run into BoundsErrors, but even with an inner constructor you are allowed to `push!` to `s.alias` and/or `s.accept`, so it would not be guaranteed that the indices are valid.