@adienes 2024-02-05
makes sense to me --- so this PR then improves safety as well as performance!