@vancleve 2024-03-04
cool, I had thought that @inbounds would improve performance, but maybe not significantly.

Seems I should close https://github.com/JuliaStats/StatsBase.jl/issues/630 then?